[AI:claude] style(clippy): clear all workspace warnings + flip CI to -D warnings (TASK-710) (#685)
* [AI:claude] style(clippy): clear all workspace warnings + flip CI to -D warnings (TASK-710)

Cleared the ~335 remaining style/pedantic clippy warnings across the
workspace and flipped the CI gate from `-W clippy::all -D
clippy::correctness` to `-D warnings`.

Scope (56 files):
- Mechanical auto-fixes via `cargo clippy --fix` (needless borrows,
  redundant closures, `.iter().any()`→`.contains()`, `.last()`→
  `.next_back()`, inline format args, derivable Default impls, manual
  `split_once`, doc-list indentation, etc.).
- Manual fixes for what `--fix` could not safely apply: 116 doc-lazy/
  overindent fixes applied by span; field-reassign→struct-update;
  manual_strip→strip_prefix; nonminimal_bool; unnecessary_filter_map→
  map; sort_by→sort_by_key(Reverse); if_same_then_else dead-branch
  removal; redundant_guards; cloned_ref→slice::from_ref; new_without_
  default; large_enum_variant boxing (PuntFlow::Terminal); type aliases
  (ReqIdData, TraceScanHit); removed two genuinely-unused fn params
  (only_used_in_recursion).
- 37 scoped `#[allow(...)]` with `// why:` justifications (no crate-wide
  allows): 20 too_many_arguments on CLI command-dispatch fns, 6
  should_implement_trait on inherent `from_str` parsers, 5
  type_complexity on local/test/private-loader tuples, 1
  large_enum_variant on the clap QueueCommand enum, 1 result_large_err
  on a tonic `Status` return, 4 dead_code on test-covered future-use
  helpers.

Correctness guard: caught and corrected `cmp_owned` auto-fixes that
stripped `PathBuf::from(...)` wrappers down to `&str`, breaking
`PathBuf == &str` comparisons (8 sites in scaffolding/main/cli tests) —
rewritten to `Path::new(...)`. Restored a BUG-96 test that a tangled
pre-existing `#[test]` duplicate had left dead. All auto-fix changes
audited as behavior-preserving.

Gates: `cargo build --release --workspace`, `cargo test --workspace`
(all crates, 0 failed), `cargo clippy --workspace --all-targets --
-D warnings` (exit 0), `cargo fmt --all -- --check` (exit 0) all pass.


* [AI:claude] ci(clippy): keep stable -D correctness gate, defer -D warnings (TASK-710)

The warning cleanup stands (335→~0 under clippy 1.92), but flipping CI to
`-D warnings` proved fragile: CI runs `dtolnay/rust-toolchain@stable` (floating),
and clippy 1.96 surfaced 23+ new lints the cleanup couldn't have seen — every
Rust release would newly-break CI with no code change. The stable, real-bug
gate is `-D clippy::correctness` (TASK-709), which already catches BUG-473's
class. Revisit full `-D warnings` only behind a pinned toolchain.


---------

+++ aida-server/src/convert.rs
@@ -383,7 +383,7 @@ pub fn custom_field_to_proto(field: &CustomFieldDefinition) -> proto::CustomFiel
         required: field.required,
         options: field.options.clone(),
         description: field.description.clone().unwrap_or_default(),
-        order: field.order as i32,
+        order: field.order,
     }
 }
